refactor: ResponseBuilder와 CgiBuilder로 응답 생성 로직 재구성
- RequestHandler → ResponseBuilder로 이름 변경
- CGI 로직을 CgiBuilder class로 분리
- ResponseFactory의 CGI 응답 생성을 CgiBuilder로 이동
